improve bind and nickname tests
